fix(bigquery): update file name
